+++ projects/oblique/src/lib/paginator/ob-paginator.module.ts
@@ -1,0 +1,14 @@
+import {NgModule} from '@angular/core';
+import {MatPaginatorIntl, MatPaginatorModule} from '@angular/material/paginator';
+
+import {ObPaginatorService} from './ob-paginator.service';
+import {obliqueProviders} from '../utilities';
+
+export {ObPaginatorService} from './ob-paginator.service';
+
+@NgModule({
+	imports: [MatPaginatorModule],
+	exports: [MatPaginatorModule],
+	providers: [ObPaginatorService, {provide: MatPaginatorIntl, useClass: ObPaginatorService}, ...obliqueProviders()]
+})
+export class ObPaginatorModule {}